Move pointer script printing logic from `Pointer` to `Pointers` class.
